requirements.txt

Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extra-fields==3.7.0
Pillow==9.2.0
